+++ static/app/components/profiling/FrameStack/frameStackContextMenu.tsx
@@ -1,0 +1,3 @@
+export function FrameStackContextMenu() {
+  return null;
+}
