+++ static/app/utils/profiling/hooks/useContextMenu.tsx
@@ -1,6 +1,29 @@
 import {useState} from 'react';
 
+import {clamp} from 'sentry/utils/profiling/colors/utils';
+import {Rect} from 'sentry/utils/profiling/gl/utils';
+
 import {useKeyboardNavigation} from './useKeyboardNavigation';
+
+export function computeBestContextMenuPosition(
+  mouse: Rect,
+  container: Rect,
+  target: Rect
+) {
+  const maxY = Math.floor(container.height - target.height);
+  const minY = container.top;
+
+  const minX = container.left;
+  const maxX = Math.floor(container.right - target.width);
+
+  // We add a tiny offset so that the menu is not directly where the user places their cursor.
+  const OFFSET = 6;
+
+  return {
+    left: clamp(mouse.x + OFFSET, minX, maxX),
+    top: clamp(mouse.y + OFFSET, minY, maxY),
+  };
+}
 
 export function useContextMenu() {
   const [open, setOpen] = useState<boolean>(false);
